refactor: create weekoverview package
