+++ app/src/main/res/layout/fragment_week_overview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WeekOverviewFragment">
+    tools:context=".weekoverview.WeekOverviewFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/weekdayListRecyclerView"
